requirements.txt
omegaconf 
opencv_python 
numpy==1.24.2 
timm 
pytorch_wavelets 
pywavelets 
pytorch_lightning 
kornia 
decord==0.6.0
einops==0.3.0
transformers==4.30.0
open_clip_torch==2.22.0
av==10.0.0